requirements.txt
ADEval==1.1.0
diffusers==0.37.0
einops==0.8.2
huggingface_hub==0.36.2
numpy==2.1.2
opencv_contrib_python==4.12.0.88
opencv_python==4.11.0.86
pandas==2.3.1
Pillow==12.2.0
scikit_learn==1.6.1
timm==1.0.25
torch==2.10.0
torchvision==0.25.0
tqdm==4.67.1
transformers==4.56.1
